docs: instructions to compile with legacy devkitPro
The project can't be compiled with the most recent
versions of the devkitPro toolchain, given the
recent refactor of the nds-related libraries. As a
temporary fix, added instructions to setup a legacy
environment through Docker to be able to compile the
app until it's updated to use the latest tools. Also
added details to help with compiling with later
versions of the C++ standard.